Improve constructor module structure
